refactor(daemon): split langfuse-trace.ts god-file into a capability-barrel module
Decompose the 2,073-line langfuse-trace.ts into a langfuse-trace/ module: a core/
foundation kernel (types, constants, redaction, primitives) plus config/, payload/,
delivery/, and reporting/ concern subdirs, each with a barrel. The root barrel
re-exports the exact prior public surface (34 names). Function bodies are moved
byte-identically; every importer is repointed to ./langfuse-trace/index.js. Every
file carries an @module docblock, every export is JSDoc'd, and a README documents
the shape.

Pure structural move, no behavior change. Capability-barrel guard registration is
deferred (that guard infra is not on main yet).
