+++ apps/daemon/src/langfuse-trace/config/index.ts
@@ -1,0 +1,5 @@
+/** @module langfuse-trace/config/index
+ * Configuration layer barrel: re-exports all environment-resolution and delivery-state
+ * utilities from the config sub-module.
+ */
+export * from './sink-config.js';
