+++ apps/daemon/src/langfuse-trace/core/index.ts
@@ -1,0 +1,8 @@
+/** @module langfuse-trace/core/index
+ * Foundation kernel barrel: shared types, tuning constants, redaction, and pure
+ * primitives; imports no sibling subdirectory.
+ */
+export * from './types.js';
+export * from './constants.js';
+export * from './primitives.js';
+export * from './redact.js';
