+++ apps/daemon/src/langfuse-trace/delivery/index.ts
@@ -1,0 +1,4 @@
+/** @module langfuse-trace/delivery/index
+ * Delivery layer barrel: re-exports all network transport functions from the delivery sub-module.
+ */
+export * from './batch-transport.js';
