+++ apps/daemon/src/langfuse-trace/index.ts
@@ -1,0 +1,43 @@
+/** @module langfuse-trace
+ * Public API for the langfuse-trace capability barrel. Re-exports only from
+ * subdirectory barrels using explicit named re-exports — never `export *`.
+ * The export surface is identical to the pre-refactor flat file: 34 names
+ * (8 values + 26 types). External code must import from this barrel only;
+ * never import directly from a subdirectory file.
+ */
+export { INPUT_MAX_BYTES } from './core/index.js';
+export type {
+  AgentEventSummary,
+  ArtifactManifestEntry,
+  ArtifactSummary,
+  AttachmentManifestEntry,
+  EventsSummary,
+  FeedbackReportContext,
+  InputTextSnapshotManifestEntry,
+  LangfuseConfig,
+  LangfuseDeliveryState,
+  LangfuseDeliveryStatus,
+  LangfuseDropReason,
+  MessageSummary,
+  ObjectManifestAccessScope,
+  ObjectManifestCompleteness,
+  ObjectManifestRetentionPolicy,
+  ObjectManifestSensitivity,
+  ObjectManifestStatus,
+  ReportContext,
+  ReportRunOpts,
+  RunSummary,
+  RuntimeInfo,
+  TelemetrySinkConfig,
+  ToolCallSummary,
+  TraceObjectSummary,
+  TraceSafeObjectManifestBase,
+  TurnInfo,
+} from './core/index.js';
+export {
+  deriveLangfuseDeliveryState,
+  readLangfuseConfig,
+  readTelemetrySinkConfig,
+} from './config/index.js';
+export { buildFeedbackPayload, buildTracePayload } from './payload/index.js';
+export { reportRunCompleted, reportRunFeedback } from './reporting/index.js';
