+++ apps/daemon/src/langfuse-trace/payload/index.ts
@@ -1,0 +1,9 @@
+/** @module langfuse-trace/payload/index
+ * Payload layer barrel: builds Langfuse ingestion batch objects — trace/span/generation/event bodies, cost, timing, diagnostics, and feedback scores.
+ */
+export * from './cost.js';
+export * from './diagnostics.js';
+export * from './feedback.js';
+export * from './summaries.js';
+export * from './timing.js';
+export * from './trace-payload.js';
