+++ apps/daemon/src/langfuse-trace/reporting/index.ts
@@ -1,0 +1,5 @@
+/** @module langfuse-trace/reporting/index
+ * Reporting layer barrel: re-exports the two public orchestration entrypoints for
+ * Langfuse trace and feedback delivery.
+ */
+export { reportRunCompleted, reportRunFeedback } from './report.js';
